@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.19'
+gem 'rails', '4.0.0'
 #added thin webserver to get rid of warning message:
 #WARN  Could not determine content-length of response body. Set content-length of the response or set Response#chunked = true
 gem 'thin'
@@ -24,6 +24,11 @@ gem 'sqlite3'
 # end
 
 # gem 'jquery-rails'
+
+group :doc do
+  # bundle exec rake doc:rails generates the API under doc/api.
+  gem 'sdoc', require: false
+end
 
 # To use ActiveModel has_secure_password
  gem 'bcrypt-ruby', '~> 3.0.1'
