@@ -2,7 +2,7 @@ class AccessController < ApplicationController
 
   layout 'admin'
 
-  before_filter :confirm_logged_in, :except => [:login, :attempt_login, :logout]
+  before_action :confirm_logged_in, :except => [:login, :attempt_login, :logout]
 
   def index
   end
