@@ -2,14 +2,14 @@ class PagesController < ApplicationController
 
   layout 'admin'
 
-  before_filter :confirm_logged_in
+  before_action :confirm_logged_in
 
   def index
     @pages = Page.sorted
   end
 
   def create
-    @page = Page.new(params[:page])
+    @page = Page.new(page_params)
     if @page.save
       flash[:notice] = "The page #{@page.name} is created"
       redirect_to(:action => 'index')
@@ -38,7 +38,7 @@ class PagesController < ApplicationController
 
   def update
     @page = Page.find(params[:id])
-    if @page.update_attributes(params[:page])
+    if @page.update_attributes(page_params)
       flash[:notice] = "The page #{@page.name} is updated"
       redirect_to(:action => 'show',:id => @page.id)
     else
@@ -57,6 +57,16 @@ class PagesController < ApplicationController
     flash[:notice] = "The page #{@page.name}has been deleted."
     redirect_to(:action => 'index')
   end
+
+  private
+
+  def page_params
+    # same as using "params[:subject]", except that it:
+    # - raises an error if :subject is not present
+    # - allows listed attributes to be mass-assigned
+    params.require(:page).permit(:name, :permalink, :position, :visible, :subject_id)
+  end
+
 end
 
 
