@@ -1,5 +1,4 @@
 class Page < ActiveRecord::Base
-  attr_accessible :name, :permalink, :position, :visible, :subject_id
 
   belongs_to :subject
   has_many :sections
@@ -9,6 +8,13 @@ class Page < ActiveRecord::Base
   validates_length_of :name, :maximum => 255
   validates_presence_of :permalink
   validates_length_of :permalink, :within => 3..255
+  # use presence_of with length_of to disallow spaces
+  validates_uniqueness_of :permalink
+  # for unique values by subject use ":scope => :subject_id"
 
-  scope :sorted, lambda{order("pages.position ASC")}
+  scope :visible, lambda { where(:visible => true) }
+  scope :invisible, lambda { where(:visible => false) }
+  scope :sorted, lambda { order("pages.position ASC") }
+  scope :newest_first, lambda { order("pages.created_at DESC")}
+
 end
